+++ packages/core/src/components/menu/VfMenu.vue
@@ -1,0 +1,15 @@
+<script setup lang="ts">
+import { useAttrs } from 'vue';
+
+defineOptions({
+  inheritAttrs: false,
+});
+
+const attrs = useAttrs();
+</script>
+
+<template>
+  <div class="vf-menu" role="menu" v-bind="attrs">
+    <slot />
+  </div>
+</template>