+++ packages/core/src/components/menu/index.ts
@@ -1,0 +1,2 @@
+export { default as VfMenu } from './VfMenu.vue';
+export { default as VfMenuItem } from './VfMenuItem.vue';