+++ packages/core/src/entries/menu.ts
@@ -1,0 +1,3 @@
+import '../styles/entries/menu.css';
+
+export { VfMenu, VfMenuItem } from '../components/menu';